requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==3.7.1
astroid==3.0.3
bcrypt==4.3.0
black==24.2.0
CacheControl==0.14.3
contourpy==1.3.0
cryptography==45.0.3
cycler==0.12.1
dill==0.4.0
distro==1.9.0
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.109.2
firebase-admin==6.9.0
fonttools==4.58.1
gcloud==0.18.3
google-ai-generativelanguage==0.6.4
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-generativeai==0.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
importlib-metadata==6.11.0
iniconfig==2.1.0
isort==5.13.2
jiter==0.10.0
jsonschema==4.22.0
jwcrypto==1.5.6
kiwisolver==1.4.7
matplotlib==3.9.4
mccabe==0.7.0
msgpack==1.1.1
mypy==1.8.0
mypy_extensions==1.1.0
numpy==1.26.3
oauth2client==4.1.3
openai==1.3.0
packaging==23.2
passlib==1.7.4
pathspec==0.12.1
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==10.2.0
platformdirs==4.3.8
pluggy==1.6.0
proto-plus==1.26.1
protobuf==4.25.3
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodome==3.23.0
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pylint==3.0.3
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
Pyrebase4==4.8.0
pytesseract==0.3.13
pytest==7.4.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.9
pytz==2023.3.post1
requests==2.31.0
requests-toolbelt==0.10.1
rsa==4.9.1
safetensors==0.4.2
sniffio==1.3.1
starlette==0.36.3
streamlit-pdf-viewer==0.0.23
tenacity==8.5.0
tokenizers==0.13.0
tomli==2.2.1
tomlkit==0.13.2
toolz==1.0.0
tqdm==4.66.1
transformers==4.33.3
typing-inspection==0.4.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.27.1
zstandard==0.22.0
# Cleaned: removed all local file paths and build artifacts. If you need a specific version, add it as 'package==version'.
